# user-cards-bridge/docs/api-documentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User Cards Bridge API",
    "description": "A comprehensive web service bridge plugin for managing user cards, roles, scheduling, and WooCommerce integration with SMS notifications.",
    "version": "1.0.0",
    "contact": {
      "name": "User Cards Bridge Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "{baseUrl}/wp-json/user-cards-bridge/v1",
      "description": "WordPress REST API endpoint",
      "variables": {
        "baseUrl": {
          "default": "https://your-wordpress-site.com",
          "description": "Your WordPress site URL"
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login user and get JWT token",
        "description": "Authenticate user credentials and return JWT token for API access",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Username or email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "User password"
                  },
                  "role": {
                    "type": "string",
                    "enum": ["company_manager", "supervisor", "agent"],
                    "description": "Optional role filter"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register new user",
        "description": "Register a new user with specified role",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "email", "password", "display_name", "role"],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Username"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "Password"
                  },
                  "display_name": {
                    "type": "string",
                    "description": "Display name"
                  },
                  "role": {
                    "type": "string",
                    "enum": ["company_manager", "supervisor", "agent"],
                    "description": "User role"
                  },
                  "supervisor_id": {
                    "type": "integer",
                    "description": "Supervisor ID (required for agents)"
                  },
                  "cards": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "description": "Card IDs (for supervisors)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Registration failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/sms/send": {
      "post": {
        "tags": ["SMS"],
        "summary": "Send SMS",
        "description": "Send SMS message via Payamak Panel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["to", "body_id", "text"],
                "properties": {
                  "to": {
                    "type": "string",
                    "description": "Phone number"
                  },
                  "body_id": {
                    "type": "string",
                    "description": "SMS body ID"
                  },
                  "text": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Text variables for SMS"
                  },
                  "customer_id": {
                    "type": "integer",
                    "description": "Customer ID (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SMS sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SMSResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "token": {
                "type": "string",
                "description": "JWT token"
              },
              "user": {
                "$ref": "#/components/schemas/User"
              },
              "role": {
                "type": "string",
                "enum": ["company_manager", "supervisor", "agent"]
              }
            }
          },
          "error": {
            "type": "null"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "User ID"
          },
          "username": {
            "type": "string",
            "description": "Username"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address"
          },
          "first_name": {
            "type": "string",
            "description": "First name"
          },
          "last_name": {
            "type": "string",
            "description": "Last name"
          },
          "display_name": {
            "type": "string",
            "description": "Display name"
          },
          "role": {
            "type": "string",
            "enum": ["company_manager", "supervisor", "agent"],
            "description": "User role"
          }
        }
      },
...
      "SMSResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "Success message"
              },
              "result": {
                "type": "object",
                "description": "SMS sending result"
              }
            }
          },
          "error": {
            "type": "null"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "data": {
            "type": "null"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code"
              },
              "message": {
                "type": "string",
                "description": "Error message"
              },
              "status": {
                "type": "integer",
                "description": "HTTP status code"
              }
            }
          }
        }
      }
    }
  }
}